backend/requirements.txt
fastapi>=0.68.0
uvicorn==0.15.0
redis==4.3.4
pydantic>=2.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
together>=1.4.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.27.0